backend/requirements.txt
FLASK==1.1
mysql-connector==2.2.9
passlib==1.7.2
pytest==5.4.3